organizando arquivos: separando a classe model do arquivo de treinamento
